==> Dataprocessing_ClusterONE/01_Yeast Biomart_InparanoidScripts.ipynb <==
### Detecting human orthologs for yeast genetic interactors of Huntingtin protein.
#### Mason etal_2013

path <- "C:/Users/Sonali/Box Sync/Huntington Interactome/MutantHTT_Interactors/Yeast/MutantHttSuppressors_green"
setwd(path)

#####################################################
### Get the required libararies
#####################################################
library("AnnotationDbi")
library("hom.Sc.inp.db")
library("biomaRt")
library ("plyr") #### just in case.

########################################
###Biomart package installation####
#######################################
source("http://bioconductor.org/biocLite.R")
biocLite("biomaRt")
source("https://bioconductor.org/biocLite.R")
biocLite("hom.Sc.inp.db")
# 
# #########################################
# ##To view documentation for the version of this package ###
# ############################################################
# browseVignettes("biomaRt")
# 
# ###################################################
# ### code chunk number 2: biomaRt
# ###################################################
# library("biomaRt")
# listMarts()
# 
# 
# 
# ###################################################
# ### code chunk number 4: listDatasets
# ###################################################
# listDatasets(ensembl)

#############################################################
### Read the file containing the gene symbols of the organism
###############################################################

yeastEGIds <- read.table(file= "R_Input_Mason_etal2013.txt", sep = "\t", header = TRUE)

# convert your Gene Symbols to ensemble gene id and ensembl protein id for your organism

yeast = useMart(host='dec2014.archive.ensembl.org',biomart = 'ENSEMBL_MART_ENSEMBL', dataset = "scerevisiae_gene_ensembl") ### specify the dataset to use 

###################################################
### filters for the yeast dataset
###################################################
filters = listFilters(yeast)
filters[1:5,]
#listFilters(yeast)

###################################################
### attributes for yeast dataset
###################################################
attributes = listAttributes(yeast)

attributes[1:5,]
#listAttributes(yeast, page="feature_page")


########################################
## Get the attributes for yeast dataset 
#######################################


rawMasonENSProtIds <- getBM(c('ensembl_gene_id', 'external_gene_name', 'ensembl_peptide_id'), filters='wikigene_name', 'external_gene_name', values= yeastEGIds$AssociatedGeneName, mart=yeast) # convert Gene Symbol to ensemble gene id and ensembl protein id for your organism
length (rawMasonENSProtIds)
dim (rawMasonENSProtIds)
class(rawMasonENSProtIds)
#rawMasonENSProtIds


#########################################
# remove duplicate values from the ensembl protein id column
#######################################

MasonENSProtIds <- rawMasonENSProtIds[!duplicated(rawMasonENSProtIds[,3]),]
#MasonENSProtIds
dim(MasonENSProtIds)
 
##############################################
# Using the Inparanoid function to find human orthologs in form of human ensembl protein IDs#
##################################################

rawHumanProtIds <- mget(unlist(MasonENSProtIds$ensembl_peptide_id) ,hom.Sc.inpHOMSA,ifnotfound=NA)

#rawHumanProtIds
class(rawHumanProtIds)
HumanProtIds <- rawHumanProtIds[!is.na(rawHumanProtIds)] ## remove all the NA values from the list
#HumanProtIds
length(HumanProtIds)

##########################################################################
## Find the ensembl gene id annd hgnc symbols for your human ensembl peptide id 
## using the getBM function and the human ensembl dataset
#######################################################################

human = useMart(host='dec2014.archive.ensembl.org',biomart = 'ENSEMBL_MART_ENSEMBL', dataset = "hsapiens_gene_ensembl")
Humanorthologs <- getBM(c('ensembl_gene_id', 'hgnc_symbol', 'ensembl_peptide_id', 'entrezgene'), filters='ensembl_peptide_id', values=HumanProtIds, mart=human) # convert Gene Symbol to ensemble gene id and ensembl protein id for your organism
Humanorthologs
dim(Humanorthologs)

## run the line below if need output in text file.

# write.table(Humanorthologs, file="MasonHumanOrthologs.txt", sep="\t", col.names = NA) ### export file in text format to the folder.


##############################################################################################
##########END ################################################################################
#############################################################################################

sessionInfo()



#### Giorgini et al_2005

path <- "C:/Users/Sonali/Box Sync/Huntington Interactome/MutantHTT_Interactors/Yeast/MutantHttSuppressors_green"
setwd(path)

#####################################################
### Get the required libararies
#####################################################
library("AnnotationDbi")
library("hom.Sc.inp.db")
library("biomaRt")
library ("plyr") #### just in case.

########################################
###Biomart package installation####
#######################################
# source("http://bioconductor.org/biocLite.R")
# biocLite("biomaRt")
# 
# #########################################
# ##To view documentation for the version of this package ###
# ############################################################
# browseVignettes("biomaRt")
# 
# ###################################################
# ### code chunk number 2: biomaRt
# ###################################################
# library("biomaRt")
# listMarts()
# 
# 
# 
# ###################################################
# ### code chunk number 4: listDatasets
# ###################################################
# listDatasets(ensembl)

#############################################################
### Read the file containing the gene symbols of the organism
###############################################################

yeastEGIds <- read.table(file= "R_Input_Giorgini_etal2005.txt", sep = "\t", header = TRUE)

# convert your Gene Symbols to ensemble gene id and ensembl protein id for your organism

yeast = useMart(host='dec2014.archive.ensembl.org',biomart = 'ENSEMBL_MART_ENSEMBL', dataset = "scerevisiae_gene_ensembl") ### specify the dataset to use 

###################################################
### filters for the yeast dataset
###################################################
filters = listFilters(yeast)
filters[1:5,]
#listFilters(yeast)

###################################################
### attributes for yeast dataset
###################################################
attributes = listAttributes(yeast)

attributes[1:5,]
#listAttributes(yeast, page="feature_page")


########################################
## Get the attributes for yeast dataset 
#######################################


rawGiorginiENSProtIds <- getBM(c('ensembl_gene_id', 'external_gene_name', 'ensembl_peptide_id'), filters='wikigene_name', 'external_gene_name', values= yeastEGIds$Suppressors, mart=yeast) # convert Gene Symbol to ensemble gene id and ensembl protein id for your organism
length (rawGiorginiENSProtIds)
dim (rawGiorginiENSProtIds)
class(rawGiorginiENSProtIds)
#rawGiorginiENSProtIds


#########################################
# remove duplicate values from the ensembl protein id column
#######################################

GiorginiENSProtIds <- rawGiorginiENSProtIds[!duplicated(rawGiorginiENSProtIds[,3]),]
#GiorginiENSProtIds
dim(GiorginiENSProtIds)
 
##############################################
# Using the Inparanoid function to find human orthologs in form of human ensembl protein IDs#
##################################################

rawHumanProtIds <- mget(unlist(GiorginiENSProtIds$ensembl_peptide_id) ,hom.Sc.inpHOMSA,ifnotfound=NA)

#rawHumanProtIds
class(rawHumanProtIds)
HumanProtIds <- rawHumanProtIds[!is.na(rawHumanProtIds)] ## remove all the NA values from the list
#HumanProtIds
length(HumanProtIds)

##########################################################################
## Find the ensembl gene id annd hgnc symbols for your human ensembl peptide id 
## using the getBM function and the human ensembl dataset
#######################################################################

human = useMart(host='dec2014.archive.ensembl.org',biomart = 'ENSEMBL_MART_ENSEMBL', dataset = "hsapiens_gene_ensembl")
Humanorthologs <- getBM(c('ensembl_gene_id', 'hgnc_symbol', 'ensembl_peptide_id', 'entrezgene'), filters='ensembl_peptide_id', values=HumanProtIds, mart=human) # convert Gene Symbol to ensemble gene id and ensembl protein id for your organism
Humanorthologs
dim(Humanorthologs)

## run the line below if need output in text file.
# write.table(Humanorthologs, file="GiorginiHumanOrthologs.txt", sep="\t", col.names = NA) ### export file in text format to the folder.


##############################################################################################
##########END ################################################################################
#############################################################################################

sessionInfo()


### Chatterjee et al 2013

path <- "C:/Users/Sonali/Box Sync/Huntington Interactome/MutantHTT_Interactors/Yeast/MutantHttSuppressors_green"
setwd(path)

#####################################################
### Get the required libararies
#####################################################
library("AnnotationDbi")
library("hom.Sc.inp.db")
library("biomaRt")
library ("plyr")

########################################
###Biomart package installation####
#######################################
# source("http://bioconductor.org/biocLite.R")
# biocLite("biomaRt")
# 
# #########################################
# ##To view documentation for the version of this package ###
# ############################################################
# browseVignettes("biomaRt")
# 
# ###################################################
# ### code chunk number 2: biomaRt
# ###################################################
# library("biomaRt")
# listMarts()
# 
# 
# 
# ###################################################
# ### code chunk number 4: listDatasets
# ###################################################
# listDatasets(ensembl)

#############################################################
### Read the file containing the gene symbols of the organism
###############################################################

yeastEGIds <- read.table(file= "R_Input_Chatterjee_etal2013.txt", sep = "\t", header = TRUE)

# convert your Gene Symbols to ensemble gene id and ensembl protein id for your organism

yeast = useMart(host='dec2014.archive.ensembl.org',biomart = 'ENSEMBL_MART_ENSEMBL', dataset = "scerevisiae_gene_ensembl") ### specify the dataset to use 

###################################################
### filters for the yeast dataset
###################################################
filters = listFilters(yeast)
filters[1:5,]
#listFilters(yeast)

###################################################
### attributes for yeast dataset
###################################################
attributes = listAttributes(yeast)

attributes[1:5,]
#listAttributes(yeast, page="feature_page")


########################################
## Get the attributes for yeast dataset 
#######################################


rawChatterjeeENSProtIds <- getBM(c('ensembl_gene_id', 'external_gene_name', 'ensembl_peptide_id'), filters='wikigene_name', 'external_gene_name', values= yeastEGIds$ConfirmedSuppressorGenes, mart=yeast) # convert Gene Symbol to ensemble gene id and ensembl protein id for your organism
length (rawChatterjeeENSProtIds)
dim (rawChatterjeeENSProtIds)
class(rawChatterjeeENSProtIds)
#rawChatterjeeENSProtIds


#########################################
# remove duplicate values from the ensembl protein id column
#######################################

ChatterjeeENSProtIds <- rawChatterjeeENSProtIds[!duplicated(rawChatterjeeENSProtIds[,3]),]
#ChatterjeeENSProtIds
dim(ChatterjeeENSProtIds)
 
##############################################
# Using the Inparanoid function to find human orthologs in form of human ensembl protein IDs#
##################################################

rawHumanProtIds <- mget(unlist(ChatterjeeENSProtIds$ensembl_peptide_id) ,hom.Sc.inpHOMSA,ifnotfound=NA)

#rawHumanProtIds
class(rawHumanProtIds)
HumanProtIds <- rawHumanProtIds[!is.na(rawHumanProtIds)] ## remove all the NA values from the list
#HumanProtIds
length(HumanProtIds)

##########################################################################
## Find the ensembl gene id annd hgnc symbols for your human ensembl peptide id 
## using the getBM function and the human ensembl dataset
#######################################################################

human = useMart(host='dec2014.archive.ensembl.org',biomart = 'ENSEMBL_MART_ENSEMBL', dataset = "hsapiens_gene_ensembl")
Humanorthologs <- getBM(c('ensembl_gene_id', 'hgnc_symbol', 'ensembl_peptide_id', 'entrezgene'), filters='ensembl_peptide_id', values=HumanProtIds, mart=human) # convert Gene Symbol to ensemble gene id and ensembl protein id for your organism
Humanorthologs
dim(Humanorthologs)

## run the line below if need output in text file.
#write.table(Humanorthologs, file="ChatterjeeHumanOrthologs.txt", sep="\t", col.names = NA) ### export file in text format to the folder.


##############################################################################################
##########END ################################################################################
#############################################################################################

sessionInfo()


#### Willingham etal_2003

path <- "C:/Users/Sonali/Box Sync/Huntington Interactome/MutantHTT_Interactors/Yeast/MutantHttEnhancers_red"
setwd(path)

#####################################################
### Get the required libararies
#####################################################
library("AnnotationDbi")
library("hom.Sc.inp.db")
library("biomaRt")
library ("plyr") #### just in case.

########################################
###Biomart package installation####
#######################################
# source("http://bioconductor.org/biocLite.R")
# biocLite("biomaRt")
# 
# #########################################
# ##To view documentation for the version of this package ###
# ############################################################
# browseVignettes("biomaRt")
# 
# ###################################################
# ### code chunk number 2: biomaRt
# ###################################################
# library("biomaRt")
# listMarts()
# 
# 
# 
# ###################################################
# ### code chunk number 4: listDatasets
# ###################################################
# listDatasets(ensembl)

#############################################################
### Read the file containing the gene symbols of the organism
###############################################################

yeastEGIds <- read.table(file= "R_Input_Willingham_etal2003.txt", sep = "\t", header = TRUE)

# convert your Gene Symbols to ensemble gene id and ensembl protein id for your organism

yeast = useMart(host='dec2014.archive.ensembl.org',biomart = 'ENSEMBL_MART_ENSEMBL', dataset = "scerevisiae_gene_ensembl") ### specify the dataset to use 

###################################################
### filters for the yeast dataset
###################################################
filters = listFilters(yeast)
filters[1:5,]
#listFilters(yeast)

###################################################
### attributes for yeast dataset
###################################################
attributes = listAttributes(yeast)

attributes[1:5,]
#listAttributes(yeast, page="feature_page")


########################################
## Get the attributes for yeast dataset 
#######################################


rawWillinghamENSProtIds <- getBM(c('ensembl_gene_id', 'external_gene_name', 'ensembl_peptide_id'), filters='wikigene_name', 'external_gene_name', values= yeastEGIds$HttEnhancergenes, mart=yeast) # convert Gene Symbol to ensemble gene id and ensembl protein id for your organism
length (rawWillinghamENSProtIds)
dim (rawWillinghamENSProtIds)
class(rawWillinghamENSProtIds)
#rawWillinghamENSProtIds


#########################################
# remove duplicate values from the ensembl protein id column
#######################################

WillinghamENSProtIds <- rawWillinghamENSProtIds[!duplicated(rawWillinghamENSProtIds[,3]),]
#WillinghamENSProtIds
dim(WillinghamENSProtIds)

##############################################
# Using the Inparanoid function to find human orthologs in form of human ensembl protein IDs#
##################################################

rawHumanProtIds <- mget(unlist(WillinghamENSProtIds$ensembl_peptide_id) ,hom.Sc.inpHOMSA,ifnotfound=NA)

#rawHumanProtIds
class(rawHumanProtIds)
HumanProtIds <- rawHumanProtIds[!is.na(rawHumanProtIds)] ## remove all the NA values from the list
#HumanProtIds
length(HumanProtIds)

##########################################################################
## Find the ensembl gene id annd hgnc symbols for your human ensembl peptide id 
## using the getBM function and the human ensembl dataset
#######################################################################

human = useMart(host='dec2014.archive.ensembl.org',biomart = 'ENSEMBL_MART_ENSEMBL', dataset = "hsapiens_gene_ensembl")
Humanorthologs <- getBM(c('ensembl_gene_id', 'hgnc_symbol', 'ensembl_peptide_id', 'entrezgene'), filters='ensembl_peptide_id', values=HumanProtIds, mart=human) # convert Gene Symbol to ensemble gene id and ensembl protein id for your organism
Humanorthologs
dim(Humanorthologs)

## run the line below if need output in text file.
#write.table(Humanorthologs, file="WillinghamHumanOrthologs.txt", sep="\t", col.names = NA) ### export file in text format to the folder.


##############################################################################################
##########END ################################################################################
#############################################################################################

sessionInfo()
