==> Dataprocessing_ClusterONE/02_Mouse Biomart_InparanoidScript.ipynb <==
### Detecting human orthologs of mouse physical interactors of Huntingtin protein.
#### Shirasaki et al_2012

path <- "C:/Users/Sonali/Box Sync/Huntington Interactome/MutantHTT_Interactors/Mouse"
setwd(path)

########################################
###Biomart and Inparanoid package installation####
#######################################
source("http://bioconductor.org/biocLite.R")
biocLite("biomaRt")
# Install Homology information for Mus musculus from Inparanoid
biocLite("hom.Mm.inp.db")

#####################################################
### Get the required libararies
#####################################################
library("AnnotationDbi")
library("hom.Mm.inp.db")
library("biomaRt")
library ("plyr") 


# #########################################
# ##To view documentation for the version of this package ###
# ############################################################
# browseVignettes("biomaRt")
# 
# ###################################################
# ### code chunk number 2: biomaRt
# ###################################################
# library("biomaRt")
#listMarts(host = 'oct2014.archive.ensembl.org')
# http://oct2014.archive.ensembl.org/index.html
# 
# 
# ###################################################
# ### code chunk number 4: listDatasets
# ###################################################
#ensembl=useMart("ensembl")
ensembl77 = useMart(host='dec2014.archive.ensembl.org', 
                    biomart='ENSEMBL_MART_ENSEMBL')
#listDatasets(ensembl77)

#############################################################
### Read the file containing the gene symbols of the organism
###############################################################

mouseGenes <- read.table(file= "R_Input_Shirasaki_etal2012.txt", sep = "\t", header = TRUE)

# convert your Gene Symbols to ensemble gene id and ensembl protein id for your organism

mouse = useMart(host='dec2014.archive.ensembl.org',biomart = 'ENSEMBL_MART_ENSEMBL' , dataset = "mmusculus_gene_ensembl") ### specify the dataset to use 

###################################################
### filters for the mouse dataset
###################################################
filters = listFilters(mouse)
filters[1:5,]
#listFilters(mouse)

###################################################
### attributes for mouse dataset
###################################################
attributes = listAttributes(mouse)

attributes[1:5,]
#listAttributes(mouse, page="feature_page")


########################################
## Get the attributes for mouse genes 
#######################################


rawShirasakiENSProtIds <- getBM(c('ensembl_gene_id', 'hgnc_symbol','ensembl_peptide_id','hgnc_id','mgi_id'), filters= 'external_gene_name', values= mouseGenes$AssociatedGeneName, mart=mouse) # convert Gene Symbol to ensemble gene id and ensembl protein id for your organism

length (rawShirasakiENSProtIds)
dim (rawShirasakiENSProtIds)
class(rawShirasakiENSProtIds)
#rawShirasakiENSProtIds

##################################################################################
### clean your data - remove any empty strings in the column of Ensembl peptide id
###################################################################################
z <- rawShirasakiENSProtIds[!rawShirasakiENSProtIds$ensembl_peptide_id == "",] # remove rows from data frame where a row matches a string 
dim(z)

#########################################
# remove duplicate values from the ensembl protein id column
#######################################
ShirasakiENSProtIds <- z[!duplicated(z[,3]),]
#ShirasakiENSProtIds
dim(ShirasakiENSProtIds)


##############################################
# Using the Inparanoid function to find human orthologs in form of human ensembl protein IDs#
##################################################

rawHumanProtIds <- mget(unlist(ShirasakiENSProtIds$ensembl_peptide_id) ,hom.Mm.inpHOMSA,ifnotfound=NA)

#rawHumanProtIds
class(rawHumanProtIds)
HumanProtIds <- rawHumanProtIds[!is.na(rawHumanProtIds)] ## remove all the NA values from the list
#HumanProtIds
length(HumanProtIds)

##########################################################################
## Find the ensembl gene id annd hgnc symbols for your human ensembl peptide id 
## using the getBM function and the human ensembl dataset
#######################################################################

human = useMart(host='dec2014.archive.ensembl.org',biomart = 'ENSEMBL_MART_ENSEMBL' , dataset = "hsapiens_gene_ensembl")
#listAttributes(human, page="feature_page")

Humanorthologs <- getBM(c('ensembl_gene_id', 'hgnc_symbol', 'ensembl_peptide_id', 'entrezgene','uniprot_swissprot'), filters='ensembl_peptide_id', values=HumanProtIds, mart=human) # convert Gene Symbol to ensemble gene id and ensembl protein id for your organism
Humanorthologs
dim(Humanorthologs)

# run the line below if you need output as a text file.
# write.table(Humanorthologs, file="ShirasakiHumanOrthologs.txt", sep="\t", col.names = NA) ### export file in text format to the folder.


##############################################################################################
##########END ################################################################################
#############################################################################################

sessionInfo()
