==> Dataprocessing_ClusterONE/03_Subsetting HIPPIE interactions_Yeast.ipynb <==
### Subsetting HIPPIE interactions with human orthologs of yeast genetic interactors of Htt protein
#### Mason et al_2013

##################################################################
## Hippie Dataset subset with Mason2013 - Human orthologs ###
##################################################################

####################################################################################
### If you have query data (df2) and want to match it with entries 
## in both columns of your original data frame (foreg. df1$A and df2$B
##############################################################################
path <- "C:/Users/Sonali/Box Sync/Huntington Interactome/MutantHTT_Interactors/Yeast/MutantHttSuppressors_green"
setwd(path)

####################################################################################
### Read the Hippie dataset ###
### example: df1 <- data.frame(A=sample(1:10, 10), B=sample(1:10, 10)) ## sample data frame 1
####################################################################################
df1 <- read.delim(file= "HIPPIE_lastupdated090514_cleaned.txt", header = TRUE, sep= "\t") ## Hippie Data set
df1[1:10,1:4] ## the first 10 rows and 4 columns of your dataframe.
# df1$ID.Interactor.A


####################################################################################
### Read the dataframe which contains the Entrez Gene IDs of the human orthologs
### example:  df2 <- data.frame(C=1:5) # sample data frame 2
####################################################################################
df2 <- read.delim(file= "MasonHumanOrthologs.txt", header = TRUE, sep= "\t") ## Data frame containing entrez gene IDs

dim(df1)
dim(df2)

########################################################################################
### Intersection function - main function doing the job!
### WILL GIVE YOU THE INTERSECTION OF THE COLUMN A WITH YOUR DATA COLUMN C FROM YOUR QUERY DATA FRAME###
### REMEMBER THIS IS A LOGICAL VECTOR.SO YOU WILL GET A BUNCH OF TRUE AND FALSE VALUES#######
##########################################################################################
selectedRows <- (df1$ID.Interactor.A %in% df2$entrezgene | df1$ID.Interactor.B %in% df2$entrezgene) ## the main function that does the work!! ###

###################################################################################################################
### Now get the resultant dataframe we are looking for with corresponding columns from your huge Hippie dataset!
##################################################################################################################

dfReduced <- df1[selectedRows,] ## dataframe containing your results.
dim(dfReduced) # dimensions of the dataframe
dfReduced[1:10,1:4] # view the first 10 and first 4 columns of the dataframe

#dfReduced ## have a look at the result.
## class(dfReduced)

# write.table(dfReduced, file="Mason2013_Hip_Interactns.txt", sep="\t", col.names = NA) ### export file in text format to the folder.

sessionInfo()




#### Giorgini et al_2005

##################################################################
## Hippie Dataset subset with Giorgini2005 - Human orthologs ###
##################################################################

####################################################################################
### If you have query data (df2) and want to match it with entries 
## in both columns of your original data frame (foreg. df1$A and df2$B
##############################################################################
path <- "C:/Users/Sonali/Box Sync/Huntington Interactome/MutantHTT_Interactors/Yeast/MutantHttSuppressors_green"
setwd(path)

####################################################################################
### Read the Hippie dataset ###
### example: df1 <- data.frame(A=sample(1:10, 10), B=sample(1:10, 10)) ## sample data frame 1
####################################################################################
df1 <- read.delim(file= "HIPPIE_lastupdated090514_cleaned.txt", header = TRUE, sep= "\t") ## Hippie Data set
df1[1:10,1:4] ## the first 10 rows and 4 columns of your dataframe.
# df1$ID.Interactor.A


####################################################################################
### Read the dataframe which contains the Entrez Gene IDs of the human orthologs
### example:  df2 <- data.frame(C=1:5) # sample data frame 2
####################################################################################
df2 <- read.delim(file= "GiorginiHumanOrthologs.txt", header = TRUE, sep= "\t") ## Data frame containing entrez gene IDs

dim(df1)
dim(df2)

########################################################################################
### Intersection function - main function doing the job!
### WILL GIVE YOU THE INTERSECTION OF THE COLUMN A WITH YOUR DATA COLUMN C FROM YOUR QUERY DATA FRAME###
### REMEMBER THIS IS A LOGICAL VECTOR.SO YOU WILL GET A BUNCH OF TRUE AND FALSE VALUES#######
##########################################################################################
selectedRows <- (df1$ID.Interactor.A %in% df2$entrezgene | df1$ID.Interactor.B %in% df2$entrezgene) ## the main function that does the work!! ###

###################################################################################################################
### Now get the resultant dataframe we are looking for with corresponding columns from your huge Hippie dataset!
##################################################################################################################

dfReduced <- df1[selectedRows,] ## dataframe containing your results.
dim(dfReduced) # dimensions of the dataframe
dfReduced[1:10,1:4] # view the first 10 and first 4 columns of the dataframe

#dfReduced ## have a look at the result.
## class(dfReduced)

#write.table(dfReduced, file="Giorgini2005_Hip_Interactns.txt", sep="\t", col.names = NA) ### export file in text format to the folder.

sessionInfo()




#### Chatterjee et al_2013

##################################################################
## Hippie Dataset subset with Chatterjee2013 - Human orthologs ###
##################################################################

####################################################################################
### If you have query data (df2) and want to match it with entries 
## in both columns of your original data frame (foreg. df1$A and df2$B
##############################################################################
path <- "C:/Users/Sonali/Box Sync/Huntington Interactome/MutantHTT_Interactors/Yeast/MutantHttSuppressors_green"
setwd(path)


####################################################################################
### Read the Hippie dataset ###
### example: df1 <- data.frame(A=sample(1:10, 10), B=sample(1:10, 10)) ## sample data frame 1
####################################################################################
df1 <- read.delim(file= "HIPPIE_lastupdated090514_cleaned.txt", header = TRUE, sep= "\t") ## Hippie Data set
df1[1:10,1:4] ## the first 10 rows and 4 columns of your dataframe.
# df1$ID.Interactor.A


####################################################################################
### Read the dataframe which contains the Entrez Gene IDs of the human orthologs
### example:  df2 <- data.frame(C=1:5) # sample data frame 2
####################################################################################
df2 <- read.delim(file= "ChatterjeeHumanOrthologs.txt", header = TRUE, sep= "\t") ## Data frame containing entrez gene IDs

dim(df1)
dim(df2)

########################################################################################
### Intersection function - main function doing the job!
### WILL GIVE YOU THE INTERSECTION OF THE COLUMN A WITH YOUR DATA COLUMN C FROM YOUR QUERY DATA FRAME###
### REMEMBER THIS IS A LOGICAL VECTOR.SO YOU WILL GET A BUNCH OF TRUE AND FALSE VALUES#######
##########################################################################################
selectedRows <- (df1$ID.Interactor.A %in% df2$entrezgene | df1$ID.Interactor.B %in% df2$entrezgene) ## the main function that does the work!! ###

###################################################################################################################
### Now get the resultant dataframe we are looking for with corresponding columns from your huge Hippie dataset!
##################################################################################################################

dfReduced <- df1[selectedRows,] ## dataframe containing your results.
dim(dfReduced) # dimensions of the dataframe
dfReduced[1:10,1:4] # view the first 10 and first 4 columns of the dataframe

#dfReduced ## have a look at the result.
## class(dfReduced)

#write.table(dfReduced, file="Chatterjee2013_Hip_Interactns.txt", sep="\t", col.names = NA) ### export file in text format to the folder.

sessionInfo()




#### Willingham etal_2003
Please note the working directory change

##################################################################
## Hippie Dataset subset with Chatterjee2013 - Human orthologs ###
##################################################################

####################################################################################
### If you have query data (df2) and want to match it with entries 
## in both columns of your original data frame (foreg. df1$A and df2$B
##############################################################################
path <- "C:/Users/Sonali/Box Sync/Huntington Interactome/MutantHTT_Interactors/Yeast/MutantHttEnhancers_red"
setwd(path)

####################################################################################
### Read the Hippie dataset ###
### example: df1 <- data.frame(A=sample(1:10, 10), B=sample(1:10, 10)) ## sample data frame 1
####################################################################################
df1 <- read.delim(file= "HIPPIE_lastupdated090514_cleaned.txt", header = TRUE, sep= "\t") ## Hippie Data set
df1[1:10,1:4] ## the first 10 rows and 4 columns of your dataframe.
# df1$ID.Interactor.A


####################################################################################
### Read the dataframe which contains the Entrez Gene IDs of the human orthologs
### example:  df2 <- data.frame(C=1:5) # sample data frame 2
####################################################################################
df2 <- read.delim(file= "WillinghamHumanOrthologs.txt", header = TRUE, sep= "\t") ## Data frame containing entrez gene IDs

dim(df1)
dim(df2)

########################################################################################
### Intersection function - main function doing the job!
### WILL GIVE YOU THE INTERSECTION OF THE COLUMN A WITH YOUR DATA COLUMN C FROM YOUR QUERY DATA FRAME###
### REMEMBER THIS IS A LOGICAL VECTOR.SO YOU WILL GET A BUNCH OF TRUE AND FALSE VALUES#######
##########################################################################################
selectedRows <- (df1$ID.Interactor.A %in% df2$entrezgene | df1$ID.Interactor.B %in% df2$entrezgene) ## the main function that does the work!! ###

###################################################################################################################
### Now get the resultant dataframe we are looking for with corresponding columns from your huge Hippie dataset!
##################################################################################################################

dfReduced <- df1[selectedRows,] ## dataframe containing your results.
dim(dfReduced) # dimensions of the dataframe
dfReduced[1:10,1:4] # view the first 10 and first 4 columns of the dataframe

#dfReduced ## have a look at the result.
## class(dfReduced)

#write.table(dfReduced, file="Willingham2003_Hip_Interactns.txt", sep="\t", col.names = NA) ### export file in text format to the folder.

sessionInfo()
