==> Dataprocessing_ClusterONE/05_Joining all HIPPIE interactions.ipynb <==
##### This script joins all the interactions obtained by subsetting the HIPPIE database with human orthologs of mouse and yeast physical/genetic interactor of Htt protein


path <- "C:/Users/Sonali/Box Sync/Huntington Interactome/MutantHTT_Interactors/Yeast/MutantHttSuppressors_green"
setwd(path)
library(dplyr) # load required library

###################################################################################
### Joining all the Hippie interaction files (mouse and yeast) to a single dataframe##
####################################################################################


df1 <- read.delim(file= "Mason2013_Hip_Interactns.txt", header = TRUE, sep= "\t") ## Mason2013 HIPPIE file (Yeast - suppressor)
dim(df1)

df2 <- read.delim(file= "Chatterjee2013_Hip_Interactns.txt", header = TRUE, sep= "\t") ## Chatterjee2013 HIPPIE file (Yeast - suppressor)
dim(df2)

df3 <- read.delim(file= "Giorgini2005_Hip_Interactns.txt", header = TRUE, sep= "\t") ## Giorgini2005 HIPPIE file (Yeast - suppressor)
dim(df3)

setwd("C:/Users/Sonali/Box Sync/Huntington Interactome/MutantHTT_Interactors/Yeast/MutantHttEnhancers_red")

df4 <- read.delim(file= "Willingham2003_Hip_Interactns.txt", header = TRUE, sep= "\t") ## Willingham2003 HIPPIE file (Yeast - enhancer)
dim(df4)

setwd("C:/Users/Sonali/Box Sync/Huntington Interactome/MutantHTT_Interactors/Mouse")
df5 <- read.delim(file= "Shirasaki_Hip_Interactns.txt", header = TRUE, sep= "\t") ## Shirasaki HIPPIE file (Mouse)
dim(df5)

#merge(df1, df2, by.x = "title1",by.y = "title1", all = TRUE)
combndf <- rbind(df1,df2,df3, df4, df5) # works ! you could insert more that 2 dataframes if needed to bind together.
dim(combndf)
# Example for merging two data frames by common field.
# Now to remove duplicate rows from the joined dataframe
finalHippieInteracns <- unique(combndf) # this gives all the unique interactions in the dataframe-- based on all columns.
dim(finalHippieInteracns)
class(finalHippieInteracns)
# finalHippieInteracns # watch your resultant dataframe

#write.table(finalHippieInteracns, file="FinalHippieInteractns.txt", sep="\t", col.names = NA) ### export file in text format to the folder.


##############################################################################################
##########END ################################################################################
#############################################################################################

sessionInfo()






### IMPORTANT NOTE: Data pre-processing.
The file "FinalHippieInteractions.txt" generated by joining all the Hippie interactions together for human orthologs of yeast and mouse interactors, was examined closely for any missing identifiers.
It was found that some of the entries in the Alt.Interactor.A and B columns did not have their Uniprot IDs.
We therefore used Uniprot database as a reference point for getting the missing uniprot ids. 
The database version, dated 6th May 2016 (UniProt Release 2016_04) was used to retreive the missing Uniprot IDs. Retrieve Id mapping tool on the website used.
The resultant modified file of Final Hippie interactions with the retreived Uniprot IDs was saved as "FinalHippieInteractions_UniprotNAsRemoved.txt".
This file will henceforth be used for processing and analysis.
