==> Dataprocessing_ClusterONE/06_Filtering HIPPIE interactions.ipynb <==
### Filtering out HIPPIE interactions with confidence score of 0.4

### IMPORTANT NOTE: Data pre-processing.
The file "FinalHippieInteractions.txt" generated by joining all the Hippie interactions together for human orthologs of yeast and mouse interactors, was examined closely for any missing identifiers.
It was found that some of the entries in the Alt.Interactor.A and B columns did not have their Uniprot IDs.
We therefore used Uniprot database as a reference point for getting the missing uniprot ids. 
The database version, dated 6th May 2016 (UniProt Release 2016_04) was used to retreive the missing Uniprot IDs. Retrieve Id mapping tool on the website used.
The resultant modified file of Final Hippie interactions with the retreived Uniprot IDs was saved as "FinalHippieInteractions_UniprotNAsRemoved.txt".
This file will henceforth be used for processing and analysis.

path <- "C:/Users/Sonali/Box Sync/Huntington Interactome/MutantHTT_Interactors"
setwd(path)


### Filter out protein interactions with confidence values less than 0.4
####################################################
y <- read.delim(file = "FinalHippieInteractions_UniprotNAsRemoved.txt", header = TRUE, sep = "\t")
y[1:5, 15:17]
dim(y)
class (y)
length(y$Confidence.Value)

threshold <- 0.4
myfilteredHippieInteractns <- subset (y,y[,17] > threshold)
myfilteredHippieInteractns1 <- myfilteredHippieInteractns[,c(5,6,17)]

notduplicated <- myfilteredHippieInteractns1[!duplicated(myfilteredHippieInteractns1),] # 860 duplicate interactions removed

uniquedf <- notduplicated[!duplicated(notduplicated[,1:2]),]

# to remove rows that have the same values in column 1 and 2 # basically removing self loops 
rows = apply(uniquedf[, 1:2], 1, function(i) length(unique(i)) > 1)
mydf <- uniquedf[rows,] # 198 self loops removed.
dim(mydf)
mydf[1:5,]
# dim of mydf is 31185 rows (interactions) and 3 columns

###########
### Export file to directory
##########################
write.table(myfilteredHippieInteractns, file = "filteredHippieInteractions_0pt3.txt",sep = "\t", col.names = NA)
write.table(mydf, file = "Clusterone_input_HippieInteractions_0pt3.txt",sep = "\t", col.names = NA)

sessionInfo()
################
### NOT RUN
### ##########
